Add files via upload
People on Low Income is a dataset from the London Datastore that reports the percentage of individuals earning less than the London Living Wage, by borough, throughout the City of London. The data is based on survey data collected by the Office for National Statistics, as part of the Annual Population Survey.